Separate thread_siblings_list on both comma and dash
On some machines, the thread siblings are listed as a range and when
executing ht-manager.pl, we get the output below:

Argument "0-1\n" isn't numeric in numeric eq (==) at ./ht-manager.pl
line 61.

Since we only care about the first digit regardless of comma or dash,
split when we encounter a dash as well.